requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
psutil==5.9.5
pytest==7.4.1
PyYAML==6.0.1
requests==2.31.0
tailer==0.4.1
tomli==2.0.1
urllib3==2.0.5
watchdog==3.0.0
